Finished basic sending/recv with clients. No names
Client will display 'You: ' when you enter a message,
but won't display other's username -- need to
implement that.

When first connecting, need to send user name along to server
to broadcast to others that you have joined.
Also need to reply to clients who are the first to join a chat.

Need to comment client.c better, and clean up server.c's as well.